feat(autoware_utils_math): split package
