+++ autoware_utils/include/autoware_utils/math/constants.hpp
@@ -1,4 +1,4 @@
-// Copyright 2020 TIER IV, Inc.
+// Copyright 2025 The Autoware Contributors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,10 +15,13 @@
 #ifndef AUTOWARE_UTILS__MATH__CONSTANTS_HPP_
 #define AUTOWARE_UTILS__MATH__CONSTANTS_HPP_
 
+#include <autoware_utils_math/constants.hpp>
+
 namespace autoware_utils
 {
-constexpr double pi = 3.14159265358979323846;  // To be replaced by std::numbers::pi in C++20
-constexpr double gravity = 9.80665;
+
+using namespace autoware_utils_math;  // NOLINT(build/namespaces)
+
 }  // namespace autoware_utils
 
 #endif  // AUTOWARE_UTILS__MATH__CONSTANTS_HPP_
