+++ autoware_utils/include/autoware_utils/math/sin_table.hpp
@@ -1,4 +1,4 @@
-// Copyright 2023 TIER IV, Inc.
+// Copyright 2025 The Autoware Contributors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,15 +15,12 @@
 #ifndef AUTOWARE_UTILS__MATH__SIN_TABLE_HPP_
 #define AUTOWARE_UTILS__MATH__SIN_TABLE_HPP_
 
-#include <cstddef>
+#include <autoware_utils_math/sin_table.hpp>
 
 namespace autoware_utils
 {
 
-constexpr size_t sin_table_size = 32769;
-constexpr size_t discrete_arcs_num_90 = 32768;
-constexpr size_t discrete_arcs_num_360 = 131072;
-extern const float g_sin_table[sin_table_size];
+using namespace autoware_utils_math;  // NOLINT(build/namespaces)
 
 }  // namespace autoware_utils
