+++ autoware_utils/include/autoware_utils/math/trigonometry.hpp
@@ -1,4 +1,4 @@
-// Copyright 2023 TIER IV, Inc.
+// Copyright 2025 The Autoware Contributors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,18 +15,12 @@
 #ifndef AUTOWARE_UTILS__MATH__TRIGONOMETRY_HPP_
 #define AUTOWARE_UTILS__MATH__TRIGONOMETRY_HPP_
 
-#include <utility>
+#include <autoware_utils_math/trigonometry.hpp>
 
 namespace autoware_utils
 {
 
-float sin(float radian);
-
-float cos(float radian);
-
-std::pair<float, float> sin_and_cos(float radian);
-
-float opencv_fast_atan2(float dy, float dx);
+using namespace autoware_utils_math;  // NOLINT(build/namespaces)
 
 }  // namespace autoware_utils
