+++ autoware_utils/package.xml
@@ -20,6 +20,7 @@
   <depend>autoware_perception_msgs</depend>
   <depend>autoware_planning_msgs</depend>
   <depend>autoware_utils_geometry</depend>
+  <depend>autoware_utils_math</depend>
   <depend>autoware_utils_pcl</depend>
   <depend>autoware_vehicle_msgs</depend>
   <depend>builtin_interfaces</depend>
