+++ autoware_utils_math/include/autoware_utils_math/constants.hpp
@@ -1,4 +1,4 @@
-// Copyright 2025 The Autoware Contributors
+// Copyright 2020 TIER IV, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,16 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE_UTILS__MATH__NORMALIZATION_HPP_
-#define AUTOWARE_UTILS__MATH__NORMALIZATION_HPP_
+#ifndef AUTOWARE_UTILS_MATH__CONSTANTS_HPP_
+#define AUTOWARE_UTILS_MATH__CONSTANTS_HPP_
 
-#include <autoware_utils_math/normalization.hpp>
-
-namespace autoware_utils
+namespace autoware_utils_math
 {
 
-using namespace autoware_utils_math;  // NOLINT(build/namespaces)
+constexpr double pi = 3.14159265358979323846;  // To be replaced by std::numbers::pi in C++20
+constexpr double gravity = 9.80665;
 
-}  // namespace autoware_utils
+}  // namespace autoware_utils_math
 
-#endif  // AUTOWARE_UTILS__MATH__NORMALIZATION_HPP_
+#endif  // AUTOWARE_UTILS_MATH__CONSTANTS_HPP_
