+++ autoware_utils_math/src/trigonometry.cpp
@@ -12,21 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware_utils/math/trigonometry.hpp"
+#include "autoware_utils_math/trigonometry.hpp"
 
-#include "autoware_utils/math/constants.hpp"
-#include "autoware_utils/math/sin_table.hpp"
+#include "autoware_utils_math/constants.hpp"
+#include "autoware_utils_math/sin_table.hpp"
 
 #include <cmath>
 #include <utility>
 
-namespace autoware_utils
+namespace autoware_utils_math
 {
 
 float sin(float radian)
 {
-  float degree =
-    radian * (180.f / static_cast<float>(autoware_utils::pi)) * (discrete_arcs_num_360 / 360.f);
+  float degree = radian * (180.f / static_cast<float>(autoware_utils_math::pi)) *
+                 (discrete_arcs_num_360 / 360.f);
   size_t idx =
     (static_cast<int>(std::round(degree)) % discrete_arcs_num_360 + discrete_arcs_num_360) %
     discrete_arcs_num_360;
@@ -47,13 +47,13 @@ float sin(float radian)
 
 float cos(float radian)
 {
-  return sin(radian + static_cast<float>(autoware_utils::pi) / 2.f);
+  return sin(radian + static_cast<float>(autoware_utils_math::pi) / 2.f);
 }
 
 std::pair<float, float> sin_and_cos(float radian)
 {
   constexpr float tmp =
-    (180.f / static_cast<float>(autoware_utils::pi)) * (discrete_arcs_num_360 / 360.f);
+    (180.f / static_cast<float>(autoware_utils_math::pi)) * (discrete_arcs_num_360 / 360.f);
   const float degree = radian * tmp;
   size_t idx =
     (static_cast<int>(std::round(degree)) % discrete_arcs_num_360 + discrete_arcs_num_360) %
@@ -78,7 +78,7 @@ std::pair<float, float> sin_and_cos(float radian)
 // subject to the license terms in the LICENSE file found in the top-level directory of this
 // distribution and at http://opencv.org/license.html.
 // The license can be found in
-// autoware_utils/third_party_licenses/opencv-license.md
+// autoware_utils_math/third_party_licenses/opencv-license.md
 // and https://github.com/opencv/opencv/blob/master/LICENSE
 
 // Modification:
@@ -86,10 +86,14 @@ std::pair<float, float> sin_and_cos(float radian)
 // 2. output of the function is changed from degrees to radians.
 namespace detail_fast_atan2
 {
-static const float atan2_p1 = 0.9997878412794807f * static_cast<float>(180) / autoware_utils::pi;
-static const float atan2_p3 = -0.3258083974640975f * static_cast<float>(180) / autoware_utils::pi;
-static const float atan2_p5 = 0.1555786518463281f * static_cast<float>(180) / autoware_utils::pi;
-static const float atan2_p7 = -0.04432655554792128f * static_cast<float>(180) / autoware_utils::pi;
+static const float atan2_p1 =
+  0.9997878412794807f * static_cast<float>(180) / autoware_utils_math::pi;
+static const float atan2_p3 =
+  -0.3258083974640975f * static_cast<float>(180) / autoware_utils_math::pi;
+static const float atan2_p5 =
+  0.1555786518463281f * static_cast<float>(180) / autoware_utils_math::pi;
+static const float atan2_p7 =
+  -0.04432655554792128f * static_cast<float>(180) / autoware_utils_math::pi;
 static const float atan2_DBL_EPSILON = 2.2204460492503131e-016f;
 }  // namespace detail_fast_atan2
 
@@ -118,8 +122,8 @@ float opencv_fast_atan2(float dy, float dx)
   if (dx < 0) a = 180.f - a;
   if (dy < 0) a = 360.f - a;
 
-  a = a * autoware_utils::pi / 180.f;
+  a = a * autoware_utils_math::pi / 180.f;
   return a;
 }
 
-}  // namespace autoware_utils
+}  // namespace autoware_utils_math
